requirements.txt
langchain==0.1.12
langchain_community==0.0.28
langchain_openai==0.0.8
langchainhub=0.1.15
pandas==2.2.1
python-dotenv==1.0.1
streamlit==1.31.1
streamlit_authenticator==0.3.1
tiktoken==0.6.0
faiss-cpu==1.8.0
pypdf==4.1.0